Simplify `gu-base` to be a collection of core tools
We don't strictly need to install a lot of the casks listed in `gu-base`, for example VLC.

Additionally, we have not standardised on any particular version managers.

In this change, we strip down `gu-base` to the core tools we use
and push the responsibility of installing applications to the user.

This will reduce friction between tools one is familiar with and the tools `gu-base` installs.